Add files via upload
Links, prices related to the components we will purchase. Please check it and let's make additions if necessary before I forward it to Cedric. By the way, there are different options available for the arms (pre-drilled and carbon fiber etc.). I'd appreciate if you guys  took a look at it.